Allow PhoneStateListeners with DEFAULT_SUBSCRIPTION
The current listener architecture does not allow the creation of
a PhoneStateListener with the DEFAULT_SUBSCRIPTION_ID. This
change will use the default subId unless one is specified. This
will cause the listener to track the current default subscription.
In addition, this will allow the creation of a listener that
will track a SIM/subscription as it is loaded/reloaded.

NOTE: This causes a minor behavior change. The default subId
tracks the default data subscription, so for registrants in
multi-SIM/subscription devices that use an explicitly set
Voice subId different from the default Data subId, a listener
previously created by this call will no longer pick the
desired subId for tracking voice-centric events. Instead a call
to getDefaultVoiceSubId() will need to retrieve the value before
the creation of the listener with an explicit subId value.

Bug:29076526
Change-Id: Iaebc76f1503edfc17265c506a3f5adde33beefe5