Basic framework support to test Avrcp/A2dp profiles.
Background and a design doc for the framework can be found in b/29097732

Add a MediaBrowserService to the test app for intercepting Media Keys.
Add a BluetoothMediaFacade for bluetooth media related test cases
Added an API to ConnectionFacade to disconnect on a specific list of profiles. Also re-factored some code here to share code.

Bug:  29097732, 28913426, 29100401
Change-Id: I1d172a1de342f19545f0a46f025effa7daf5ad4a
(cherry picked from commit 901bd7709544583e9497dc970105093180a5debf)