Add a Persistent Listener for Connectivity Change Broadcasts
Toggling airplane mode triggers bluetooth to toggle on and off.
Because this is not an explicit toggle of bluetooth, we need a
way to listen for the adapter state change without explicitly
toggling bluetooth. Adding a way to start and stop listening to
bluetooth state change broadcasts achieves this goal.

Bug: 31473466
Test: none
Change-Id: I694fe1d070c05db3af61a9163662faa6a652bd38