[sl4a]WifiManagerFacade: Fix connect logic.
WifiManagerFacade.wifiConnect() should not be adding, enabling and
re-connecting to the network. Fixed the connect logic by passing
the config and directly invoking the connect() API, which should
take care of doing other necessary steps before connecting.

Also changed the method name to wifiConnectByConfig(), to
distinguish it from connecting to a network by passing the NetId.

BUG: 31755596
TEST: All WifiManagerTest.py tests passed

Change-Id: I46d65eca62b81e811e8a70d6d801144b95136302